Current prototype of game with board and controls
Currently made it so player works as a boolean with basically an
on/off switch to make the marker be X and O, and as of right now,
it can accurately get the right winner. However, the code is very messy
and as it is right now, it is very difficult to find a way to add new
functions.